consumer/requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
python-dotenv==1.0.0
pytest==6.2.5
pytest-cov==4.1.0